v2.8.2 (#12)
- Xcode 11 support
- Improved Mac Catalyst support
- Use Network.framework for reachability on iOS 12+
- Expose `shouldUseExtendedBackgroundIdleMode`
- Disable `connectivityOptions` on `TNLRequestConfiguration` for iOS 13.0 (ok on iOS 13.1+)
- Add _expensive_ and _constrained_ conditions to the reachability flags on iOS 13+
- Add new iOS 13 metadata from transaction metrics
- Use __FILE_NAME__ instead of __FILE__
- Optimize `tnl_safeSubdataNoCopyWithRange:`
- Optimize NSDictionary `tnl_copyToMutable:uppercase:`
- Fix unreliable temporary file writing for across app runs

+++ TNLExample/TNLXLotsOfRequestsViewController.m
@@ -371,7 +371,7 @@ static const BOOL kUseThumbnail = NO;
                 delay++;
             }
         } else {
-            _progressView.backgroundColor = [UIColor colorWithRed:0.75 green:0.1 blue:0.1 alpha:0.0];
+            _progressView.backgroundColor = [UIColor colorWithRed:(CGFloat)0.75 green:(CGFloat)0.1 blue:(CGFloat)0.1 alpha:(CGFloat)0.0];
         }
     } else {
         NSString *tempPath = [NSTemporaryDirectory() stringByAppendingPathComponent:[[NSUUID UUID] UUIDString]];
